feat: support under javascript disabled
