+++ src/components/Hero.astro
@@ -10,7 +10,6 @@ const { github, zenn, x } = siteConfig;
   <h1>oriverk.dev</h1>
   <p><span class="text-gradient">Agr.</span> → ? / Bicycle</p>
   <div class="links">
-    <slot />
     <a href="/blog" title="Blog" data-astro-prefetch>
       <Icon type="pencil" size="medium" />
       <span class="sr-only">Blog link</span>
